requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LangChain and OpenAI
langchain>=0.1.0,<0.2.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-core>=0.1.0

# Vector database
chromadb==0.4.22

# Document processing
pypdf==3.17.4

# Data validation
pydantic==2.5.2

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# OpenAI SDK
openai>=1.0.0,<2.0.0
tiktoken>=0.5.0

# Additional dependencies
numpy>=1.24.0
typing-extensions>=4.5.0